Move s3ng related test helpers into s3ng
